requirements.txt
Flask
PyMuPDF
Pillow
PyPDF2
reportlab
